feat: formal verification integration for contract validation (#889)

+++ backend/api/src/cache.rs
@@ -230,7 +230,7 @@ impl CacheLayer {
             let key = format!("abi:{}", contract_id);
             let mut conn = cm.clone();
             if let Err(e) = conn
-                .set_ex::<_, _, ()>(&key, &abi, self.config.abi_ttl_secs as usize)
+                .set_ex::<_, _, ()>(&key, &abi, self.config.abi_ttl_secs as u64)
                 .await
             {
                 tracing::warn!("Redis put_abi error: {}", e);
@@ -387,7 +387,7 @@ impl CacheLayer {
             let rkey = format!("contracts:{}", key);
             let mut conn = cm.clone();
             if let Err(e) = conn
-                .set_ex::<_, _, ()>(&rkey, &value, self.config.metadata_ttl_secs as usize)
+                .set_ex::<_, _, ()>(&rkey, &value, self.config.metadata_ttl_secs as u64)
                 .await
             {
                 tracing::warn!("Redis put_contracts error: {}", e);
@@ -469,7 +469,7 @@ impl CacheLayer {
         if let Some(cm) = &self.redis_cm {
             let mut conn = cm.clone();
             if let Err(e) = conn
-                .set_ex::<_, _, ()>(&ns_key, &value, self.config.metadata_ttl_secs as usize)
+                .set_ex::<_, _, ()>(&ns_key, &value, self.config.metadata_ttl_secs as u64)
                 .await
             {
                 tracing::warn!("Redis put_contract_meta error: {}", e);
@@ -539,7 +539,7 @@ impl CacheLayer {
         if let Some(cm) = &self.redis_cm {
             let mut conn = cm.clone();
             if let Err(e) = conn
-                .set_ex::<_, _, ()>(&ns_key, &value, self.config.search_ttl_secs as usize)
+                .set_ex::<_, _, ()>(&ns_key, &value, self.config.search_ttl_secs as u64)
                 .await
             {
                 tracing::warn!("Redis put_search error: {}", e);
@@ -593,7 +593,7 @@ impl CacheLayer {
         if let Some(cm) = &self.redis_cm {
             let mut conn = cm.clone();
             if let Err(e) = conn
-                .set_ex::<_, _, ()>(&ns_key, &value, self.config.stats_ttl_secs as usize)
+                .set_ex::<_, _, ()>(&ns_key, &value, self.config.stats_ttl_secs as u64)
                 .await
             {
                 tracing::warn!("Redis put_stats error: {}", e);
